+++ src/fico4omnet/buffer/can/CanBuffer.cc
@@ -53,7 +53,7 @@ void CanBuffer::putFrame(cMessage* msg){
     frames.push_back(frame);
     emit(queueLengthSignal, static_cast<unsigned long>(frames.size()));
     queueSize+=static_cast<size_t>(frame->getByteLength());
-    emit(queueSizeSignal, queueSize);
+    emit(queueSizeSignal, static_cast<unsigned long>(queueSize));
 }
 
 void CanBuffer::deleteFrame(unsigned int canID) {
@@ -62,7 +62,7 @@ void CanBuffer::deleteFrame(unsigned int canID) {
     frames.remove(tmp);
     emit(queueLengthSignal, static_cast<unsigned long>(frames.size()));
     queueSize-=static_cast<size_t>(tmp->getByteLength());
-    emit(queueSizeSignal, queueSize);
+    emit(queueSizeSignal, static_cast<unsigned long>(queueSize));
     delete tmp;
 }
 
@@ -71,7 +71,7 @@ void CanBuffer::deleteFrame(CanDataFrame* frame) {
     frames.remove(frame);
     emit(queueLengthSignal, static_cast<unsigned long>(frames.size()));
     queueSize-=static_cast<size_t>(frame->getByteLength());
-    emit(queueSizeSignal, queueSize);
+    emit(queueSizeSignal, static_cast<unsigned long>(queueSize));
     delete frame;
 }
