+++ src/fico4omnet/buffer/flexray/FROutputBuffer.cc
@@ -32,7 +32,7 @@ void FROutputBuffer::putFrame(cMessage* msg) {
     frames.push_back(frame);
     emit(queueLengthSignal, static_cast<unsigned long>(frames.size()));
     queueSize+=static_cast<size_t>(frame->getByteLength());
-    emit(queueSizeSignal, queueSize);
+    emit(queueSizeSignal, static_cast<unsigned long>(queueSize));
 }
 
 void FROutputBuffer::sendingCompleted(int id) {
