+++ src/fico4omnet/bus/can/CanBusLogic.cc
@@ -171,7 +171,7 @@ void CanBusLogic::sendingCompleted() {
         ids.erase(eraseids.at(it));
         delete *(eraseids.at(it));
     }
-    emit(arbitrationLengthSignal,ids.size());
+    emit(arbitrationLengthSignal, static_cast<unsigned long>(ids.size()));
     eraseids.clear();
     errored = false;
     if (scheduledDataFrame != NULL) {
@@ -223,7 +223,7 @@ void CanBusLogic::registerForArbitration(unsigned int canID, cModule *module,
     Enter_Method_Silent
     ();
     ids.push_back(new CanID(canID, module, signInTime, rtr));
-    emit(arbitrationLengthSignal,ids.size());
+    emit(arbitrationLengthSignal, static_cast<unsigned long>(ids.size()));
     if (idle) {
         cMessage *self = new cMessage("idle_signin");
         scheduleAt(simTime() + (1 / (bandwidth)), self);
@@ -245,7 +245,7 @@ void CanBusLogic::checkoutFromArbitration(unsigned int canID) {
             break;
         }
     }
-    emit(arbitrationLengthSignal,ids.size());
+    emit(arbitrationLengthSignal, static_cast<unsigned long>(ids.size()));
 }
 
 void CanBusLogic::colorBusy() {
